feat: Car class 파일 분리
